+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 This file is used to list changes made in each version of the Dnsmasq cookbooks.
 
+## Unreleased
+
+- Fix systemd-resolved conflict (fixes #35)
+
 ## v0.3.1
 
 - Switch testing from serverspec to inspec